app/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
psycopg2-binary==2.9.9
pika==1.3.2
pydantic==2.4.2 